admin page now uses non database data

+++ src/lib/Database/delete.ts
@@ -1,7 +1,7 @@
-import { connectToDb, db } from "./surreal";
+import { connectToDb, database, db } from "./surreal";
 
 export async function deleteItemFromDatabase(id: string){
-    console.log('deleting ' + id);
+    if(!database) return;
     await connectToDb();
     await db.delete(id); 
 }
